Add missing merge clauses in DeltaTableWriter (#46)
DeltaTableWriter provides the option to configure the writer in MERGE
mode using the output_mode_params field by providing in it a list of
merge clauses in the merge_builder key. This is especially useful in
case the Delta table or the DataFrame to be merged are not available
upfront. Not all merge clauses were supported: the merge clause provided
was being validated against a list of valid clauses which did not
include whenMatchedUpdateAll and whenNotMatchedInsertAll clauses. As
part of this PR:

- validation of the clauses now happens in the output_mode_params field
validator (currently this is happening only when the merge is being
executed)
- the list of available merge clauses is sourced directly from available
methods in DeltaMergeBuilder class and now includes all of them

## Related Issue
https://github.com/Nike-Inc/koheesio/issues/43

## Motivation and Context
Now it's possible to use all merge clauses when configuring the
DeltaTableWriter

---------

Co-authored-by: riacom_nike <riccardo.iacomini@nike.com>