[FEATURE] Enhance HttpStep to support bearer token and improve header handling (#158)
Adds support for including a bearer token in the HTTP module headers along with proper masking to ensure the token's 'secret' status is maintained.

## Changes
- Added default ability to include a bearer token in the headers.
- Implemented masking for the bearer token in any output to maintain its 'secret' status.
- Updated the get_headers method in the Http module to accommodate these changes.
- Added unit tests

## Related Issue
Closes #157 

## Motivation and Context
The current implementation of get_headers did not support adding a
bearer token by default and did not properly mask the token in the
output. This posed a security risk. This PR addresses these issues by
adding default support for bearer tokens and ensuring that they are
masked in any output.

---------

Co-authored-by: Danny Meijer <10511979+dannymeijer@users.noreply.github.com>