[FEATURE] Support for Delta table history (#163)
- Added `describe_history()` method to `DeltaTableStep`, enabling
fetching of Delta table history as a Spark DataFrame.
- Added `is_date_stale()` method to assess if data in a specified table
is stale based on defined time intervals or a specific intended refresh
day.

## Motivation and Context
It allows one to get a Delta table's history (based on Delta Log) as a
Spark DataFrame. It also provides means for checking the staleness of
data within Delta tables based on defined time intervals and specific
weekdays designated for refreshing.