bugfix - PR #179 - *Delta > Writer*: Fixed checking merge builder type in DeltaTableWriter
There was a small error in validating merge builder type when using the `merge_builder` param with `DeltaTableWriter`. It required that the value passed was either a `list` or a `delta.tables.DeltaMergeBuilder`
instance so validation failed when an instance of `delta.connect.tables.DeltaMergeBuilder` was passed.

## Motivation and Context
One cannot pass a correct DeltaMergeBuilder instance as a merge builder
for `DeltaTableWriter` if it's not a `delta.tables.DeltaMergeBuilder`
instance or a `list`. Connect merge builders should also be supported.

## How Has This Been Tested?
4 dedicated test cases have been added to test the merge builder
validation.