bugfix - PR #180 - *Snowflake > Spark* Ensure matched inserts sync to snowflake
Ensure a new INSERT is synchronized to Snowflake, even when a match is already found.

## Motivation and Context
When a delta table is VACUUMed before a sync to Snowflake was triggered, a updated record will be considered as an "INSERT" record. However the target table might already contain this record. This PR ensures that this record is still updated as would be desired.

## How Has This Been Tested?
This was a bug in our code. This code currently lives in our repo and fixed the issue.

---------

Co-authored-by: Bruno De Deken <bruno.dedeken@nike.com>