Peek transformation (#182)
adds a Peek transformation that can be used during development and debugging of tasks.

## Motivation and Context
It can be handy to inspect the schema / peek the content of the data
during the development / debugging, more so if the data pipeline
involves multiple transformation. The peek transformation can be
inserted in the most critical sections to easily inspect the content or
schema, and can be easily enabled / disabled using a flag so that it
does not need to be removed from the pipeline when that is moved to a
production environment.